Allow different formation orientations
The program will now test formations in the four orientations revolving
around the Minecraft positive Y axis.
This change quadruples the search time but is more leniant for the user
inputing the formation blocks.

This does not attempt to mirror the formation. So the axis used by the
user must remain consistent with Minecraft's coordinate space when
allowing for rotations only.